backend: hwcomposer: allow scheduling an output destruction after the current frame event
When reacting on hotplug events, we might trigger an output destruction
while the compositor is still accessing it.

Allow scheduling destruction after the current frame event via the new
wlr_hwcomposer_output_schedule_destroy.

This works by setting the should_destroy flag and scheduling another
frame.
If the flag is set, the signal_frame() callback will trigger output
destruction.

This shouldn't be racy since the only other place where we schedule a
new frame is during commits - so in the worst case the timer would be
rearmed (but since the flag is set the callback would still destroy
the output).

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>