libinput: avoid going through wlr_session to manage input devices
A session/seat manager is usually used to manage input devices. It's a neat
way to manage multi-seat environments and ensures that the input
devices file descriptors are muted once the session goes inactive.

This happens on wlroots as well, and in Droidian this reflects with the usage
of systemd-logind to manage input devices.

A pain point in mobile devices is the high frequency of sleep state changes.
This in turn results in a session looping between active/inactive multiple
times in a span of seconds. Every loop means multiple DBus calls (in the
logind case) to take/release input devices.

These events end up queuing up on the compositor, delaying its availability
or in the worst case, crashing it.

Avoid all of this by managing input devices directly. This is equivalent
to the 'noop' session backend (but we don't block). There will be only
one file descriptor per-input device, and no take/release calls.

One downside is that we lose multi-seat functionality, but we don't target
that anyway. Obiously we assume that the user running the compositor can
also read and write on the input devices.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>